requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
streamlit==1.26.0
underthesea==6.6.0
transformers==4.33.1
torch==2.0.1
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.9
chromadb==0.4.10
sentence-transformers==2.2.2
scikit-learn==1.3.0
seqeval==1.2.2
datasets==2.14.5
stanza==1.8.2
accelerate>=0.21.0
pytest>=7.4.0
python-dotenv==1.0.0
httpx>=0.24.1
pypdf==4.1.0
python-docx==1.1.0
python-multipart==0.0.9
google-genai>=1.71.0